add simple tests with fock classes
